kube context
the flags are now --kubecontext and --kubeconfig

helm drops --kube-context as it is expected to be assumed by the tunnel it provides while we disable the helm provided tunnel, so this is a helm bug. We may adapt to this in the future by using the helm provided tunnel when it is provided, however we work around this by using --kubecontext which is passed directly to Barrelman in all cases.

--kubeconfig adapts to its environment and uses helms provided KUBE_CONFIG when available and obeys --kubeconfig when it is seen by Barrelman.

The result is both switches work inside our outside helm plugin environment.