Adds unit testing for apply command, 44% coverage.
Tests revealed a bug in Apply command. There is a closure around the upgrade and install commands in order to implement retry logic for Install. This logic overloaded the use of "err" when the retries had been exceeded which resulted in no error being reported. A separate value which is global to the closure is now being used to hold the error state which resolves the bug and is proven in the new tests.