Adds File and Dir handling
data.source.type can now be configured with "file"  or "dir" in addition to the previously available "git" option.

"file" uses the data.source.location as the filepath to a fully formed tgz chart file.

"dir" uses the data.source.location as the source to build a chart archive and will resolve dependencies as defined in the manifest.

All three, "dir", "file", and "git" can be mixed in the same manifest, however "file" does not participate in dependency resolution.