Add delete release fallback on Install failure
Install failures can occur because the target release has been installed since the start of Barrelman. The new state would normally require either an "upgrade" or "force install (delete)" in order to succeed.  This change effectuates a "force install" when a failure to install has occurred. This could be enhanced to filter on specific causes, however we do not yet know what those are.
Overall this change results in the desired behavior.