Migrate to go modules, final
This PR includes configuration for Go modules with vendoring enabled.

Go modules has become a practical requirement because we still require vendoring due to restricted access to depended packages and we are now importing Barrelman as a package in Bootswain. Dep vendoring leaves vendored packages in place which results in core packages being initialized twice whereas Go modules does not have this behavior.